PoC: Allow any TimeZone implementation
Make the TimeZone parametrised in as many places as possible. This
allows using any TimeZone implementations for most of the logic. Most
notably, this allows using `chrono::Local`.

Using `chrono::Local` makes it much easier  to construct TimeZone
instances from parsed `VTIMEZONE` components and to interoperate with
other crates.

Some notes:

- There's a few new `.clone()` calls. This is because `chrono_tz::Tz`
  implements `Copy`, but other `TimeZone` implementation do not. In
  reality, these values were already being copied before anyway, but
  this was done implicitly.
- For error messages, `to_string()` cannot be used, since `TimeZone`
  does not implement `Display`. The implementations in the `chrono`
  package are the most obvious offenders here. `Debug` is now used
  instead for error messages.
- The `validate_until` validation is no longer called. It assumes that
  all `RRules` use the `rrule::Tz` TimeZone implementation. I don't see
  an obvious way to fix it, and this is still a pending item.
- `VALIDATION_PIPELINE` and `FILTERS` were defined in a way that breaks
  reusability with generic parameters, so these were dropped in favour
  of just referring to them in-line.

This is a proof of concept (which I wrote on a long train trip). Tests
pass, but this probably warrants some extra discussion.

Fixes: https://github.com/fmeringdal/rust-rrule/issues/24