simplify RRuleSetIter
The RRuleSetIter will peek all the rrule iterators to find the earliest
date available. And peek the rdates list. And then consume the date which
is the earliest by either stepping the corresponding rrule iterator or
popping from the rdate list.

This makes it possible to skip the rrule "queue" and manual buffering.